change test file for controller test
